requirements.txt
python-dotenv==0.19.0
requests==2.26.0
qrcode[pil]==7.3.1
flask==2.0.2
cryptography==3.3.2
python-telegram-bot==20.0a0
